requirements.txt
django==2.2.8
djangorestframework==3.9.2
django-cors-headers==3.2.0
channels==2.4.0
channels-redis==2.4.1